refactor: rename package repository -> persistence, Query -> 영속성 패키지로 이동
