+++ src/main/java/com/meoguri/linkocean/domain/bookmark/service/ReactionServiceImpl.java
@@ -8,10 +8,11 @@ import org.springframework.transaction.annotation.Transactional;
 import com.meoguri.linkocean.domain.bookmark.entity.Bookmark;
 import com.meoguri.linkocean.domain.bookmark.entity.Reaction;
 import com.meoguri.linkocean.domain.bookmark.entity.Reaction.ReactionType;
-import com.meoguri.linkocean.domain.bookmark.repository.ReactionRepository;
+import com.meoguri.linkocean.domain.bookmark.persistence.GetBookmarkByIdQuery;
+import com.meoguri.linkocean.domain.bookmark.persistence.ReactionRepository;
 import com.meoguri.linkocean.domain.bookmark.service.dto.ReactionCommand;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
-import com.meoguri.linkocean.domain.profile.service.GetProfileByUserIdQuery;
+import com.meoguri.linkocean.domain.profile.persistence.GetProfileByUserIdQuery;
 
 import lombok.RequiredArgsConstructor;
