+++ src/main/java/com/meoguri/linkocean/domain/linkmetadata/persistence/LinkMetadataRepository.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.linkmetadata.repository;
+package com.meoguri.linkocean.domain.linkmetadata.persistence;
 
 import java.util.Optional;
