+++ src/test/java/com/meoguri/linkocean/domain/linkmetadata/persistence/LinkMetadataRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.linkmetadata.repository;
+package com.meoguri.linkocean.domain.linkmetadata.persistence;
 
 import static org.assertj.core.api.Assertions.*;
